WalletBuilders Coin File
I searched online and found this company called WalletBuilders, they give you the option of a "FREE" cryptographic currency builder where it generates a fully working node package to get you up and running. Well everything works as follows, but as you're allowing the core program to update with the peer network it loses connection to the server. I don't know what went wrong but everything checked out, but the initial understanding of how they give instructions didn't give the response of the finding the right RPC port which is set by default from the package content after it created.